+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Version 1.0 introduces bundling for semi-fungible and fungible assets, serialize
 - `OpenSeaPort::computeFees()` now takes in either the `fees` from a collection or a single asset, and returns a new and more consistent type, `ComputedFees`
 - `schemaName` has been moved out of main method calls and into the `Asset` type
 - `OpenSeaAPI::getAsset()` now accepts named arguments, consistent with other methods
+- `OpenSeaPort::createFactorySellOrders` now accepts an `assets` parameter consistent with the other `Asset` type parameters in other methods
 - `OpenSeaPort::getTokenBalance()` no longer defaults to the WETH address if no `tokenAddress` is set. And it accepts a `schemaName` parameter instead of an ABI
 - `OpenSeaPort::approveNonFungibleToken()` has been renamed to `approveSemiOrNonFungibleToken()`, though it always worked for semi-fungible tokens like ERC-1155